+++ src/main/java/com/github/pjfanning/xlsx/impl/StreamingWorkbookReader.java
@@ -90,7 +90,9 @@ public class StreamingWorkbookReader implements Iterable<Sheet>, AutoCloseable {
     } catch(IOException e) {
       throw new ReadException("Unable to read input stream", e);
     } catch(RuntimeException e) {
-      f.delete();
+      if (f != null) {
+        f.delete();
+      }
       throw e;
     }
   }
